Update issue #94 Status: Needs-Documenting
If no UnitOfWork is active upon dispatching an event, and one is active when the event has been dispatched, that UnitOfWork is either committed, or rolled back, depending of the outcome of command handling. This will cause an ImplicitUnitOfWork to be committed, even when repository.save() is not called.
If a unit of work was already active, it is expected to be managed by another instance, and remains untouched.
Note that the command bus will not start a UnitOfWork. It will only commit one is one was created (and not cleaned up) when dispatching a command.