Upcasters can now convert one object to an arbitrary number of objects.
This feature allows Upcasters to upcast an event into zero, one or multiple
events. Previously it was only possible to upcast an event to exactly one
other event.

The upcaster interface as well as the serializer interface have been changed
to support this feature. Both deserialization and upcast methods now return
a list of objects. The same goes for upcasting and deserialization of types,
where each element in the list of types should correspond with the element on
the same index in the list of deserialized objects.

Lazy deserialized objects now have an extra index field that points to the
index of the deserialized object on the list of deserialized objects. Meaning
that multiple lazy deserialized can have a pointer to the same serialized
object but will, upon deserialization, return different objects.