Upcasting is now a task of the event store instead of the serializer.
Upcasting is now a task of the event store which makes the serializer concept more loosely coupled
and the serializer's interface more easier to use. To accomplish this the LazyDeserializingObject now
contains a LazyUpcastingObject. The LazyUpcastingObject is responsible for upcasting a given serialized
object only once and only when getObject is called on it.

The LazyDeserializingObject keeps track of which part of the upcasted object it is interested in de-serializing.
Meaning that each entry in a set of upcasted events that all have the same original event are wrapped in a
LazyDeserializingObject, all these wrappers point to a different index on the same LazyUpcastingObject.