Fixed potential GC race condition issue in Saga
The SagaCache holds a weak reference to saga entries. However, a transaction holding a SagaEntry will not prevent the cached sage from being cleaned up. When cleaned up, it may be reconstructed without seeing the uncommitted changes held by the EntityManager. SagaEntry now holds a (hard) reference to the Saga being stored, to prevent cleanup while the transaction is running.