Fixed deadlock possibility in SagaManager
This deadlock could appear when a Saga dispatches a command which in turn
publishes an event that invokes the same saga. The code in the Saga is
executed using a lock on the saga. If the command cannot be executed directly
in the UoW, the lock on the saga is released. If another thread is doing a
similar operation, it may hang on the aggregate's lock, while keeping the
lock on the saga. This issue is resolved by making sure a lock is held while
the thread participates in an active UoW.