Unit of work listeners are now handled in a specific order of precedence.
Introduced a new class which sole purpose it is to handle the order of
precedence in which unit of work listeners are called during the specific
stages of a unit of work. For example during the prepare commit stage the
listeners are handled in the order they are registered in the unit of work.
While during the after commit stage the order of precedence is reversed.
This change was introduced to further refine the behaviour in which unit of
work listeners are expected to function. This is handy, for example, when
using the AuditingListener which can perform audits before any transaction is
started and should only write an audit entry when the transaction has already
been successfully committed.