Introduced batch sending of events as described in issue #39.
Extracted from the EventBus javadoc:
The events that are submitted to the event bus in a batch should be
semantically coherent (the batch has to make sense as a whole in the
application's domain). The Event Bus can then treat them as such. This is
especially useful if you want to have transactional behaviour: dispatch the
whole batch, or to dispatch no events in the batch at all (when something goes
wrong).