Improved CommandHandlerPrefetcher for DisruptorCommandBus
It now uses weak references to keep track of loaded aggregates. Aggregates that
are still used in the disruptor cannot be garbage collected. Aggregates that
have not been used for at least one cycle are eligible for collection. This
prevents out of memory exceptions when loading large numbers of aggregates.