Added cache configuration to DisruptorCommandBus
It is now possible to use a cache to configure the TTL of aggregates in the
disruptor command bus. If an Aggregate is not available in the disruptor's
buffer, it will attempt to load it from the cache. If it is not in the cache
either, it is initialized using events from the event store.