Issue #41: Added the ability to do upcasting to the FileSystemEventStore.
A FileSystemEventStore now contains an UpcasterChain to be able to perform
upcasting after reading an event from the filesystem. To facilitate this feature
the generic EventReader and Writer have been replaced with a
FilesystemReader/Writer. The FileSystemReader is particularly useful since
it allows the EventStore to do upcasting easily by not wrapping its result in a
DomainEventMessage directly like the EventReader does. Instead it returns
the raw data from the filesystem as a FileSystemEventEntry.