Fixes issue #52. onPrepareCommit doesn't get all events to be published
The UnitOfWorkListener.onPrepareCommit method was not given the entire list of
events to be published. Instead, only the events that had been directly
registered for publication through UnitOfWork.publishEvent we in the list. The
fix ensures that all events that will be published on any EventBus on UoW
commit are contained in the eventsToPublish list.

See also: https://github.com/AxonFramework/AxonFramework/issues/52