Refactoring of snapshot event structure
Instead of storing an AggregateSnapshot object with an aggregate insize,
snapshots are now just the aggregate on itself. Furthermore, the
AggregateFactory is now responsible for recognizing snapshots and initiailizing
the aggregate that is found inside. The AbstractAggregateFactory helps in the
implementation.
Finally, the SpringPrototypeAggregateFactory will properly post-process
aggregates from a snapshot using the bean definition of the protype bean.
Fixes issue #60