Reverted the state synchronization logic in JGroupsCommandBusConnector
Just to make sure that all members are always properly synchronized, they start
with a getState, after which each member sends updates. This way, if the
channel was already connected, but did not have a receiver configured yet, it
will always end up in the correct state.

Issue #AXON-20 Comment