Added support for DispatchInterceptors to all CommandBus implementations
The AsyncronousCommandBus, SimpleCommandBus and DisruptorCommandBus now also
have support for DispatchInterceptors. These interceptors are always invoked in
the thread that delivers a Command to the Command Bus, allowing it to attach
context information to the command that is stored as ThreadLocals, as is often
the case with Security Contexts.

Issue #AXON-24 Fixed