Modification in the way AMQP configuration is provided
Instead of having only a single configuration for all clusters, it is now
possible to provide configuration on a per-cluster basis. Values that have not
been explicitly set will automatically revert to defaults set on the terminal.
The Cluster MetaData property to set the configuration on is "AMQP.Config".

Issue #AXON-28 Tag Requires Testing