Extracted the revision resolution to separate interface
Instead of having a protected method on the serializer, there is now an option
to provide a RevisionResolver to resolve the revision of a serialized object.
Three implementations are provided: AnnotationRevisionResolver, which checks
for the @Revision annotation on the payload, SerialVersionUIDRevisionResolver,
which uses the serialVersionUID field, and FixedValueRevisionResolver, which
as you may guess, returns a fixed value.

Issue #AXON-31 Fixed