Added better support for CommandGateway in Saga Test Fixture
It is now possible to register a gateway by providing its interface. The
gateway is registered as a resource as well. Optionally, you may also provide a
stub implementation of the gateway to define behavior. By doing so, you have
the ease of use of a mocking framework, combined with the ability to verify
dispatched commands in the fixture.

Issue #AXON-39 Fixed