Added pure hibernate implementation of DefaultEventEntryStore
The JPAEventStore will check whether Hibernate is on the classpath. If so, it
will use the Hiberante specific implementation. Otherwise, it falls back to a
pure JPA implementation, that doesn't have the same performance
characteristics, though. Especially when reading more events than can fit in a
single batch.

Issue #AXON-40 Fixed