Refactoring of Saga Repository and Manager to imprve transaction support
The Saga repository doesn't need to cache Saga instances anymore for any other
purpose than performance. The SagaCache has been moved to the SagaManager,
which is now responsible for dealing with thread-safety and Saga concurrency.

API changes were needed, but are limited to the SagaRepository and some
(abstract) implementations.

Issue #AXON-44 Fixed