The concept of a Unit of Work is now more tightly coupled with transactions
UnitOfWork is now "transaction aware", meaning that it can be bound to a
transaction. When that is the case, an additional UnitOfWorkListener method
will be invoked: onPrepareTransactionCommit, which allows logic to be executed
after storage and publication of events, but just prior to the commit of a
transaction.

This also means that the (Spring)TransactionInterceptor is deprecated. Instead,
one should set the TransactionManager directly on the Command Bus.

Issue #AXON-53 Fixed