Added wrong identifier detection in test fixtures
When using the wrong identifier to load an aggregate in the Test Fixtures, Axon
will now throw an exception. This can occur when the command contains an
identifier that does not match the identifier used in the "given" events.

This also uncovered a problem in the LockingRepository, where the identifier
used to load an aggregate was used to obtain a lock, while that actual
identifier of the aggregate was used to release the lock. When identifiers
don't match, this would lead to exceptions and leakage of locks.

Issue #AXON-61 Fixed