ParameterResolverFactory in tests is more flexible and safer
The ParameterResolverFactory used by the test fixtures is now aware of the fact
that multiple tests may be running in the same JVM. It will only inject
resources in the test that registered the resources. Furthermore, it only
requires resources that are actually required for the test to be injected.

Issue #AXON-100 Fixed
Issue #AXON-101 Fixed