Annotations on interface methods are also inspected
The @CommandHandler and @EventHandler annotations may now also be places on an
interface implemented by the handler. Methods on the interface are evaluated
after the methods declared on the class implementing the interface, but before
the methods on the class' superclass. So if class A extends B and implements C,
the order in which methods are inspected is A, C, B.

Issue #AXON-136 Fixed