Removed self-subscription from annotation adapters
Self-subscription cause problems in cases where objects were proxied. In that
case, the proxy was circumvented, as the inner object would register itself
with the infrastructure components.
In the new situation, object expose the state necessary to properly subscribe
them. The Spring components will automatically subscribe the components during
the startup phase of the ApplicationContext (instead of during the bean post
processing phase, which was currently the case).

Issue #AXON-157 Fixed