Fixed issue with nested UoW and event ordering
The order of events were wrong when using 3 layers of nested events. When using
a cache, the events were also out of order when appending to the event store.
This is solved by registering nested UoW with the root UoW instead. The root
UoW maintains event publication order.

Issue #AXON-180 Fixed