Fixed NPE issue in JGroupsConnector
A NPE would occur in cases where a node sends a join message and disconnects
immediately. The receivers of the join message cannot map the origin of a
message to a node. The NPE did not affect behavior, but looked ugly in the log
files. Now, a nice log message is written.

Issue #AXON-182 Fixed