Added support for custom ParameterResolver for aggregates and sagas
Aggregates and Sagas are not (always) constructed in the control of Axon.
Therefore, it is not possible to inject a ParameterResolverFactory. For those
cases, there is now a AnnotationConfiguration class where one can register the
ParameterResolverFactory to use for a specific annotated type (or for all
types).

Issue #AXON-123 Fixed