CachingEventSourcingRepository no longer needs to reload on version conflict
The CachingEventSourcingRepository would force a full reload of an aggregate
when the actual version did not match the expected version. Now, it will use
the cached aggregate, and load the "unseen" events from the event store, if the
event store implements the PartialEventStreamSupport interface. Otherwise,
behavior is unchanged.

Issue #AXON-150 Fixed