Added ability to attach resources to a Unit of Work
This feature will be used in upcoming versions to manage resources used by
building block implementations, such as JPA Event Store or AMQPTerminal to
ensure the same resource can be used during the processing of a Unit of Work.
This prevents the waste of creating new resources several times.

Issue #AXON-200 Fixed