Reverted to default to the ClasspathParameterResolverFactory.
All components allow a custom ParameterResolverFactory to be configured. B
default, the ClasspathParameterResolverFactory is used, which is configured
to use the class loader of the class to resolve parameters for.

Spring beans can no longer be injected into the Event Handlers of aggregates.
Although it seems a limitation, it can be seen as a bad practice to require
any type of resources in event sourcing handlers.

Issue #AXON-123 Fixed