Added ExtendedUpcaster interface, which allow for extra flexibility
The UpcasterChain implementations will recognize this interface and invoke the
upcast(SerializedType, SerializedObject) method instead of the
upcast(SerializedType). This allows the implementation to decide the type based
on the SerializedObject's contents

In the next major release, this method will be included in the Upcaster
interface itself.

Issue #AXON-220 Fixed