Added support for serialized data type customization in Event Stores
The JPA Event Store and (to some extent) the JDBC event store lacked the
ability to change the column type of serialized data. They both expected a
byte[] (or at least a binary column type).
This commit makes it possible to completely change the column type of the
serialized payload and meta data of events.

Issue #AXON-253 Fixed