Changed API to use Cluster as first level citizen on EventBus
While previously a relatively hidden feature of the ClusteringEventBus, the
Cluster is now a first level component in the Axon infrastructure. The Cluster
is an important component in the infrastructure, as it defines the distribution
characteristics of the application.

Note that when using Spring, it is not mandatory to define a cluster in the
application context. When component scanning for beans with Event Handlers,
Axon will create a default cluster if none is explicitly defined.

Issue #AXON-307 Fixed