AsynchronousCluster uses UnitOfWork to time Event publication
Instead of queueing an Event directly, the AsynchronousCluster will register a
callback on the Unit of Work to queue the event when the Unit of Work has been
committed. This prevents a potential situation where an event is being
processed before it is committed to the Event Store, or when a Unit of Work is
rolled back, an Event was processed which was not present in the Event Store.

Issue #AXON-322 Fixed