Changed Unit of Work API
The UnitOfWork is unaware of the exact type of work to be done. This makes it
more usable for activities performed by Async Event Handlers, Saga Managers,
etc.
The code compiles, but many tests fail. These tests will be easier to fix as
soon as some other API's have also changed (EventStore acting as a Bus; support
for direct publication to Event Store / Bus, which will then assign the task to
the proper phase of the Unit of Work; etc.)