Replace the default event backlog with a thread-safe implementation during replays
If events are being replayed on a cluster in a live environment it is possible that the event backlog is modified while an other thread is iterating over it. This raised a ConcurrentModificationException because the default backlog queue we used wasn't thread-safe. This has been fixed by replacing the queue with a ConcurrentLinkedQueue.

Issue #AXON-338 Fixed