Make LockManager more generic and fix OptimisticLockManager
LockManagers are no longer exclusive for the LockingRepository but have been moved to common. They are used to manage locks on shared resources like aggregates and sagas the same way the IdentifierBasedLock used to. The OptimisticLockManager now effectively locks on validateLock() instead of obtainLock(). LockManagers return an AutoCloseable Lock when asked to provide a lock.