Add interceptor that starts and commits Transactions
When the unit of work is committed the interceptor also commits the transaction. Same for rollback. I also changed the API of the TransactionManager: when it is asked to start a new transaction the manager returns a Transaction object. Committing or rolling back of the transaction is done on the returned object.