AsyncSagaManager closes transactions, even when no sagas were processed
The AsyncSagaManager would only commit a transaction if any sagas were affected
during a batch. This caused a number of transactions to remain open. In certain
database systems, this would prevent maintenance tasks to be executed.
This commit contains a fix, which will always close any database transactions,
so that the application is in a clean state when idle.

Issue #AXON-5157 Fixed