Enhance MetaData Api
Added a public 'and(String key, Object value)' method to add a key-value pair to a MetaData object instead of implementing a complete Builder class, since this supports a simple builder construction like MetaData.with(...).and(...).and(...) etc whilst not having the overhead of an additional class.
The static method to instantiate a MetaData with a single key-value pair was already present.

Issue #AXON-364 Fixed