Sagas no longer need to extend a base class
Similar to the Aggregates, Sagas are now simple obects with annotations on their methods. SagaEventHandler annotated methods are regular MessageHandlers which are wrapped for saga specific behavior.

The API's are designed and implementations are working. Javadoc still needs some tuning. At least the API's allow for the Spring support to be creatd.

Issues #AXON-304 Comment