Create event store tracking EventProcessor
Events can now be supplied to an EventProcessor in two ways: either by subscribing the processor directly to the event bus or by streaming events from the event store and supplying those events to the event processor. The event processor then passes the events on to an EventProcessingStrategy which may decide to handle the events directly or asynchronously. Finally, when the events are handled they are passed to registered EventListeners.