Improved snapshotting API
The SnapshotTriggerDefinition is now configured explicitly on the EventSourcingRepository. Besides a NoOp implementation, there is an implementation that is triggered based on the number of events handled by an aggregate. The state of the trigger is attached to the aggregate instance, meaning it is cached alongside it, making configuration much easier.