Expose association values of managed sagas that haven't been persisted yet
If a saga is being operated on in one thread and during that operation a new association is made, that association should be visible to other threads. If not it is possible that a saga will not receive events that it should have. Furthermore, with this commit access to a saga is locked while the saga is being operated on.