AXON-14202 fixed: read last event sequence number from the upstream event entry stream
This fixes an issue for aggregates that receive an event stream where the last event has a sequence number that is lower than that of the last entry in the event store which can happen as result of upcasting. If the aggregate would publish an additional event then that would receive an old sequence number.