AXON-29701 fixed: process events with the same tracking token in the same batch
If an event entry was upcast to more than one event there was a risk that a TrackingEventProcessor would not handle all events (e.g. in case of an application shutdown after the first event). To work around this issue TrackingEventProcessors now always handle events with the same tracking token in a single batch.