Use key in the Unit of Work to store Monitors
Currently, Processors rely on the message of the unit of work as a key to a map of Monitors. However, the message can be modified during processing (unitOfWork.transformMessage), causing the Monitor to be unreachable.
Instead, the monitor is now stored as a resource in the Unit of Work, so that it can be retrieved during cleanup.