Changed UnitOfWork API to loosen the contract on transformMessage
The current API was to trict, and didn't allow for a tranformMessage(m -> m.andMetaData(...)) structure when the UoW declared a wildcard type. The reason is that the andMetaData method cannot guarantee that the returned implementation is exactly the same. In fact, it often isn't.
The API is loosened to allow any type of mesage as return value, but it is cast to the type the Unit of Work expects (usualy CommandMessage or EventMessage). Message Transformers should take caution not to return an invalid implementation.