Split JGroups and Spring Cloud distributed CommandBus into separate modules
Split the JGroups specific and Spring Cloud specific implementation of the CommandRouter and CommandBusConnector into two separate modules.
This will help spring starter projects to auto-implement the distributed implementation based on which of both is found on the classpath.
Following to this, all the corresponding distributed CommandBus code has to be moved to axon-core/commandhandling/distributed to have a single point of entrance for the interfaces e.g.