Added TransactionManager to database backed EventStorageEngine implementations
Some database engines do not allow the use of blob fields outside of the scope of a transaction. The EmbeddedEventStore may use a dedicated thread to read events from the store and make them available to the processors. That thread does not explicitly start a transaction, causing problems with these database engines. Starting a transaction (if none already exists) when reading a batch of events, solves this issue.