Added concise logging of failures when using commandGateway.send(cmd)
The send method with the command as only parameter returns a CompletableFuture that should be queried for the result of the command. However, not doing so will cause any errors to pass by unnoticed. To prevent multiple stacktraces for the same error, only a summary of the failure is logged. For the error details (stack trace and causes), the CompletableFuture must be checked.