Added bean ordering support in Spring Configuration
Beans are sorted using AnnotationAwareOrderComparator before they are registered with the AxonConfiguration. This comparator allows beans to be annotated with @Order and @Priority to define their relative order.

Support for deprecated "OrderResolver" has been entirely removed. When using the ConfigurationAPI, handlers are invoked in the order they are registered with the Configuration instance.