Added method to extend the claim on token
This method call fetchToken by default, but allows implementations of the TokenStore to provide a more efficient implementation for extending the claim on a token, without changing the token itself.
The JPATokenStore provides this implementation, executing a query instead of persisting an updated TokenEntry. The latter caused the Token itself to be persisted as well, even if the value didn't change.

Addresses the issue described in #260