Prevent deadlocks in async saga processor when using row locks
The asynchronous saga manager processes incoming events in batches for
efficiency. This can result in the same set of sagas being loaded in
different orders depending on the particular events queued up at any
given time.

In a clustered environment, if the saga repository acquires exclusive
locks on saga objects when it loads them, this can result in deadlocks
when two nodes attempt to load the same sagas in different orders at
the same time.

It would be possible to optimistically try loading all the sagas and
then recover if a deadlock was encountered, but this turns out to be
messier than it looks. Some databases (e.g., PostgreSQL) require the
current transaction to be rolled back or restored to a savepoint when
an error such as a deadlock is encountered, and at this level of the
code there's no clean way to do that in a storage-layer-independent
way. Also, on some databases there is a deadlock detection delay which
would make the optimistic approach significantly slower if deadlocks
happen with much frequency.

This change takes a more pessimistic approach: imposing a strict order
on the sagas that are loaded in a given UnitOfWork. The order is dumb
but fair: it will refuse to load a saga whose ID is less than the ID
of any already-loaded saga, or in other words, it requires sagas to be
loaded in ascending ID order. When an event arrives that would violate
that order, the current UnitOfWork is committed, which (if a transaction
manager is active) releases any previously-acquired database locks.

The downside to this approach is that, worst case, it increases the
number of database writes by a factor of the Disruptor batch size.
This could represent a significant performance hit if the sagas involved
are all very large and costly to read and write. But that worst case
would require pretty degenerate application behavior and in empirical
testing on an application of moderate complexity, the number of
UnitOfWork commits increased by around 15% and the wall-time penalty
was a few percent, more than outweighed by the ability to freely process
events for the same saga on multiple nodes.