TrackingProcessor should wait for shutdown to complete
Without the wait, a shutdown procedure might shut down resources needed by the event processor (such as the EntityManager), which causes exception when processing events taken in just prior to the shutdown.

In upcoming versions, a solution allowing for parallel shutdown of several processors should be considered. However, this requires an API change that should not be done in a bugfix release.

Resolve issue #383