Specify a sequence name for axon entities' IDs
This makes the default strategy, AUTO, explicit, and names the generated
sequence to avoid conflicts with other strategies which also use the
default name, 'hibernate_sequence'.

For example, if one uses an H2 database, without specifying a
@SequenceGenerator, it chooses a NoopOptimizer resulting in a sequence
named 'hibernate_sequence' with an increment of 1. If a domain entity
does specify a @SequenceGenerator, without a unique name, it will try to
define 'hibernate_sequence' with the default increment of 50, and
validation will fail with "HibernateException: Multiple references to
database sequence [hibernate_sequence] were encountered attempting toset
conflicting values for 'increment size'.  Found [1] and [50]"