Further partitioning of membership updates
Further partitioning of membership updates, by splitting the
getRemoteMembershipInformation into a metadata check.
If it contains the expected membership information metadata, it will be
retrieved from there. Otherwise, it delegates a call to the
membershipInformationFromNonMetadataSource, which throws an
UnsupportedOperationException for the default SpringCloudCommandRouter
implementation.
Additionally, the buildSimpleMember() now retrieves the local instance
URI it self instead of getting it passed through, and is made package
private for reuse.

[#350]