Refactoring of TrackingEventProcessor
It now uses a ThreadFactory to spawn threads for processing. As long as
there are Segments left to claim and the number of threads hasn't
reached the configured maximum, more threads are spawned. The Thread
responsible for spawning worker thread will pick up the last Segment.

This commit also contains a bit of cleaning up things I came across.