Fixed possible illegal stealing of token in use
It was possible that a token in use by a processor was stolen bij
another processor, if the processing of an event exceeded the timeout
of a token.
This has been fixed by forcing an exclusive lock on the token (in the
underlying database), preventing another node from accessing it. Once
access is granted, the token should have been updated to a more recent
timestamp.

Resolves #422