Fix inner unit of work rollback issue on parent rollback
This commit fixes an issue where a middle unit of work is rolled back,
but the rollback handlers of an inner unit of work weren't invoked. Now,
on commit, each unit of work will register a rollback handler with the
parent to invoke its own rollback handlers in that case.

Resolves #421