Use the ForwardingMode to route events
Use the ForwardingMode to route events. There is a cut off if
ForwardingMode. == NONE to not forward any events. On ALL the entity
receives all the events within this entity. And lastly if the mode is
ROUTING_KEY, the routing value of the event is matched to the entity
identifier.
The javadoc has ben adjusted correctly to reflect this.

[#388]