Introduce test class for entity event routing
Introduce test class for entity event routing separated from the
ModelInspector test class to decrease file size.
Tests a single entity, list entity and map entity aggregate, as
well as an entity with ForwardingMode set to NONE and an entity with a
specific routing key different from the entity id.

[#388]