Fixed potential race condition when quickly updating node membership
When multiple updates of a node are sent in quick succession, it is
likely that they are received out of order by the recipient.
This commit adds an index to each message, allowing the recipient to
detect whether a message is superceded by another. These messages are
then ignored, as only the last message needs to be processed.

Resolves #424