Created stub implementation of AggregateLifecycle that can be used in tests
Especially in unit tests of aggregates that do not use event sourcing.
When they call apply(), and exception is thrown indicating that no
lifecycle has been registered. The StubAggregateLifecycle can be used
to register a lifecycle that just keeps track of applied events.

In JUnit tests, the StubAggregateLifecycleRule can be used to
automatically start these lifecycles during tests.

Resolves #336