Improved use of the AggregateModel in repositories
It is now possible to provide the AggregateModel, which paves the way
for different ways of defining aggregates than using Annotations.
Repositories will honor these models, but revert to the default
annotation based versions.

The Axon Configuration API uses an AggregateMetaModelFactory instance
obtained from the Configuration to create models for aggregates.