Allow state stored Aggregates to emit DomainEventEntries
When the GenericJpaRepository is configured with an Event Store, it will
configure new Aggregate instances to emit DomainEventMessages, as well
as Aggregates that have been loaded and already have sequenced events
in the Event Store. Aggregates that had already been created, and do
not have any events in the event store that can be traced back to it,
will keep on publishing regular events.

Relates to issue #373