Serializers are now configurable through application.properties
A notion of a Message Serializer has been added, which is used to
serialize Command, Query and Event (if no Event Serializer is specified)
payload and meta data.

When using Spring Boot, application.properties can be used to define the
type of serializer to use for events, message and generally. The most
specific suitable serializer is autowired into each of the components.

Resolves #446