Resolves issue with staged event ordering
In a situation where an Aggregate is evicted from first-level cache,
and the aggregate is reconstructed using stored and staged events from
multiples layers of nested units of work, the order in which the staged
events were applied may have been incorrect. This commit fixes that
issue, ensuring events are applied in the order they were originally
published by the units of work.

Resolves #551