AutoConfigured Repository instances are now available as beans
Repository instances created for `@Aggregate` annotated aggregates
weren't made available as beans in the application context. Therefore,
it was not possible to simply autowire a repository as a bean.
Repositories were accessible using
`Configuration.repository(MyAggregate.class)`, but that would require
additional logic to be able to autowire them in scanned components.

Resolves #620