Failure to start a Transaction will roll back unit of work
This change will ensure a Unit of Work is rolled back when a transaction
fails to start. Previously, there were cases in which a Unit of Work
would "linger". This caused problems when threads were reused to handle
more messages.

This commit is inspired by pull requests #617 and #618, which addressed
the problem described in issue #616.

Resolves #616